+++ src/lib.rs
@@ -87,7 +87,6 @@ pub mod advisories;
 pub mod bans;
 mod cfg;
 pub mod diag;
-mod index;
 /// Configuration and logic for checking crate licenses
 pub mod licenses;
 pub mod sources;
